Refactor automatic segmentation v2 API design
